# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "Users_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the current user’s profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "Users_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update the current user’s profile",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Permanently deletes the account and all associated data (sessions, conversations, searches). Requires the current password.",
        "operationId": "Users_deleteAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Account deleted"
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete account",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "description": "Requires the current password. All other sessions are signed out.",
        "operationId": "Users_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT API",
    "description": "\nREST API powering the **EchoGPT** multi-AI chat Chrome extension.\n\n### Authentication\n1. `POST /api/v1/auth/register` or `POST /api/v1/auth/login` returns an **access token** (short lived JWT) and a **refresh token** (opaque, long lived, rotated on every use).\n2. Send the access token as `Authorization: Bearer <token>`. Click **Authorize** above to use it here.\n3. When the access token expires call `POST /api/v1/auth/refresh`. Re-using an already rotated refresh token revokes the whole session.\n\n### Errors\nEvery error uses the same envelope: `{ statusCode, error, message, details?, path, timestamp, requestId }`.\nThe `x-request-id` response header echoes the correlation id for support/debugging.\n\n### Quotas\nAI endpoints (chat, AI-assisted search) consume the caller's daily plan quota. Exceeding it returns **429**.\n",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "3f6c1a2e-8a41-4a57-9d0f-1b8c0f2e6a11"
          },
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "fullName": {
            "type": "string",
            "nullable": true,
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://cdn.example.com/avatar.png"
          },
          "role": {
            "enum": [
              "ADMIN",
              "USER"
            ],
            "type": "string",
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "lastLoginAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "avatarUrl",
          "role",
          "isActive",
          "emailVerified",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Field level details, present for validation errors.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-29T10:15:30.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "6f1c2c1e-6c47-4f0e-9a36-0c5f0f4b1a2d"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 2048,
            "format": "uri",
            "example": "https://cdn.example.com/avatar.png",
            "description": "HTTPS URL of the avatar image. Send null to remove it."
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "maxLength": 128,
            "example": "Str0ngPass!"
          },
          "newPassword": {
            "type": "string",
            "description": "New password. 8-128 characters, at least one letter and one number.",
            "example": "N3wStr0ngPass!",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "DeleteAccountDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 128,
            "description": "Current password, required to confirm deletion",
            "example": "Str0ngPass!"
          }
        },
        "required": [
          "password"
        ]
      },
...
    }
  }
}